requirements.txt
langchain==0.1.16
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
anthropic
langchain-core
duckduckgo-search